[fix] create user

+++ tutorial2/pkg/view/user.go
@@ -1,6 +1,6 @@
 package view
 
 type User struct {
-	ID       uint16 `json:"id"`
+	ID       int    `json:"id"`
 	Username string `json:"username"`
 }
